+++ cnd/src/config/mod.rs
@@ -3,13 +3,9 @@ mod serde_bitcoin_network;
 mod serde_duration;
 mod settings;
 
-use config as config_rs;
 use libp2p::Multiaddr;
-use std::{
-    ffi::OsStr,
-    net::IpAddr,
-    path::{Path, PathBuf},
-};
+use serde::{Deserialize, Serialize};
+use std::{net::IpAddr, path::PathBuf};
 
 pub use self::{
     file::File,
@@ -19,18 +15,6 @@ pub use self::{
 #[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
 pub struct Database {
     pub sqlite: PathBuf,
-}
-
-use serde::{Deserialize, Serialize};
-
-impl File {
-    pub fn read<D: AsRef<OsStr>>(config_file: D) -> Result<Self, config_rs::ConfigError> {
-        let config_file = Path::new(&config_file);
-
-        let mut config = config_rs::Config::new();
-        config.merge(config_rs::File::from(config_file))?;
-        config.try_into()
-    }
 }
 
 #[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
